LayoutTests/fast/images/text-recognition/mac/image-overlay-maintain-selection-during-size-change-expected.txt

PASS selectionWidth is 200
PASS selectionHeight is 200
PASS selectionWidth became 50
PASS selectionHeight is 50
PASS selectionWidth became 200
PASS selectionHeight is 200
PASS successfullyParsed is true

TEST COMPLETE
